refactor(app): Fleet TUI Plan 4 — delete screens.py, wire HermiaApp
- Delete src/hermia/screens.py (398-line single-host TUI, replaced by hermia.tui)
- Rewire app.py: boot HermiaApp instead of EvalApp, drop --host arg,
  move RESULTS_DIR import from hermia.screens → hermia.submit
- Delete tests/unit/test_screens_pilot.py (patched hermia.screens.*)
- Delete tests/unit/test_screens.py; extract scoring tests to test_scoring.py
- Fix test_fleet.py: patch hermia.submit.RESULTS_DIR, drop EvalApp mock
