+++ src/hermia/app.py
@@ -1,4 +1,4 @@
-"""Hermia EvalApp — entry point."""
+"""Hermia entry point."""
 
 import argparse
 import os
@@ -6,27 +6,8 @@ import sys
 from datetime import date
 from pathlib import Path
 
-from textual.app import App
-
 from hermia import __version__
-from hermia.metrics import detect_gpu
-from hermia.runner import detect_mode, get_available_models
-from hermia.screens import SelectionScreen
-
-
-class EvalApp(App):  # type: ignore[type-arg]
-    TITLE = "Hermia LLM Eval"
-    BINDINGS = [("q", "quit", "Quit")]
-
-    def __init__(self, fleet_mode: bool = False, repeat: int = 1) -> None:
-        super().__init__()
-        self.fleet_mode = fleet_mode
-        self.repeat = repeat
-        self.model_list = get_available_models()
-        self.gpu_info = detect_gpu()
-
-    def on_mount(self) -> None:
-        self.push_screen(SelectionScreen(repeat=self.repeat))
+from hermia.tui.app import HermiaApp
 
 
 def _positive_int(value: str) -> int:
@@ -39,11 +20,6 @@ def _positive_int(value: str) -> int:
 def main() -> None:
     parser = argparse.ArgumentParser(description="Hermia LLM Eval")
     parser.add_argument("--version", action="version", version=f"%(prog)s {__version__}")
-    parser.add_argument(
-        "--host",
-        default="http://localhost:11434",
-        help="Ollama base URL (default: http://localhost:11434)",
-    )
     parser.add_argument(
         "--repeat",
         type=_positive_int,
@@ -109,7 +85,7 @@ def main() -> None:
 
     if args.audit is not None:
         from hermia.audit import run_audit
-        from hermia.screens import RESULTS_DIR
+        from hermia.submit import RESULTS_DIR
 
         source = RESULTS_DIR if args.audit is True else Path(args.audit)
         if not source.exists():
@@ -125,8 +101,8 @@ def main() -> None:
     if args.fleet:
         from hermia.fleet import load_fleet_config, run_fleet
         from hermia.results import load_jsonl
-        from hermia.screens import RESULTS_DIR
         from hermia.sink.submission import SubmissionSink
+        from hermia.submit import RESULTS_DIR
 
         verbosity = 1 if args.verbose else (-1 if args.quiet else 0)
         try:
@@ -169,10 +145,7 @@ def main() -> None:
 
         sys.exit(0)
 
-    os.environ["HERMIA_HOST"] = args.host.rstrip("/")
-    fleet_mode = detect_mode(args.host) == "fleet"
-
-    EvalApp(fleet_mode=fleet_mode, repeat=args.repeat).run()
+    HermiaApp().run()
 
 
 if __name__ == "__main__":
