--- tests/unit/test_screens.py
@@ -1,156 +1,0 @@
-"""Unit tests for screens.py — pure functions and scoring logic."""
-
-import socket
-
-from hermia.scoring import backfill_aggregates as _backfill_aggregates
-from hermia.scoring import compute_scores as _compute_scores
-from hermia.screens import (
-    RunnerScreen,
-    _resolve_fleet_host,
-    _sanitize_model_id,
-)
-
-# ── _sanitize_model_id ────────────────────────────────────────────────────────
-
-def test_sanitize_replaces_colon() -> None:
-    assert _sanitize_model_id("llama3:8b") == "llama3_8b"
-
-
-def test_sanitize_replaces_dot() -> None:
-    assert _sanitize_model_id("llama3.1:8b") == "llama3_1_8b"
-
-
-def test_sanitize_replaces_both() -> None:
-    assert _sanitize_model_id("qwen2.5:32b-instruct") == "qwen2_5_32b-instruct"
-
-
-def test_sanitize_clean_name_unchanged() -> None:
-    assert _sanitize_model_id("mistral") == "mistral"
-
-
-# ── _compute_scores ───────────────────────────────────────────────────────────
-
-def _result(
-    model: str, json_valid: bool, schema_compliant: bool, tps: float = 50.0
-) -> dict[str, object]:
-    return {
-        "model": model,
-        "json_valid": json_valid,
-        "schema_compliant": schema_compliant,
-        "tokens_per_sec": tps,
-    }
-
-
-def test_compute_scores_all_pass() -> None:
-    results = [_result("m1", True, True), _result("m1", True, True)]
-    scored = _compute_scores(results)
-    assert len(scored) == 1
-    model, jp, sp, ag, tps = scored[0]
-    assert model == "m1"
-    assert jp == 1.0
-    assert sp == 1.0
-    assert ag == 1.0
-
-
-def test_compute_scores_all_fail() -> None:
-    results = [_result("m1", False, False)]
-    _, jp, sp, ag, _ = _compute_scores(results)[0]
-    assert jp == 0.0
-    assert sp == 0.0
-    assert ag == 0.0
-
-
-def test_compute_scores_agentic_formula() -> None:
-    # json_valid=True, schema_compliant=False → jp=1.0, sp=0.0, ag=0.4
-    results = [_result("m1", True, False)]
-    _, jp, sp, ag, _ = _compute_scores(results)[0]
-    assert jp == 1.0
-    assert sp == 0.0
-    assert abs(ag - 0.4) < 1e-9
-
-
-def test_compute_scores_partial_schema_pass() -> None:
-    results = [
-        _result("m1", True, True),
-        _result("m1", True, False),
-    ]
-    _, jp, sp, ag, _ = _compute_scores(results)[0]
-    assert jp == 1.0
-    assert sp == 0.5
-    assert abs(ag - (1.0 * 0.40 + 0.5 * 0.60)) < 1e-9
-
-
-def test_compute_scores_avg_tps() -> None:
-    results = [_result("m1", True, True, tps=40.0), _result("m1", True, True, tps=60.0)]
-    _, _, _, _, tps = _compute_scores(results)[0]
-    assert tps == 50.0
-
-
-def test_compute_scores_sorted_by_agentic_descending() -> None:
-    results = [
-        _result("weak", False, False),
-        _result("strong", True, True),
-        _result("mid", True, False),
-    ]
-    scored = _compute_scores(results)
-    models = [s[0] for s in scored]
-    assert models == ["strong", "mid", "weak"]
-
-
-def test_compute_scores_multiple_models_separate() -> None:
-    results = [_result("a", True, True), _result("b", False, False)]
-    scored = _compute_scores(results)
-    assert len(scored) == 2
-    assert scored[0][0] == "a"
-    assert scored[1][0] == "b"
-
-
-# ── _backfill_aggregates ──────────────────────────────────────────────────────
-
-def test_backfill_aggregates_empty_list() -> None:
-    """Test that _backfill_aggregates handles empty list without error."""
-    _backfill_aggregates([])  # Should not raise
-
-
-# ── RunnerScreen init ─────────────────────────────────────────────────────────
-
-def test_runner_screen_stores_models_and_tests() -> None:
-    screen = RunnerScreen(["qwen2.5:32b", "llama3:8b"], ["tool-calling-basic"])
-    assert screen.models == ["qwen2.5:32b", "llama3:8b"]
-    assert screen.test_ids == ["tool-calling-basic"]
-    assert screen.all_results == []
-
-
-# ── _resolve_fleet_host ───────────────────────────────────────────────────────
-
-def test_resolve_fleet_host_dns_success(monkeypatch) -> None:
-    _resolve_fleet_host.cache_clear()
-    monkeypatch.setattr(socket, "gethostbyaddr", lambda ip: ("node1", [], []))
-    url, hostname = _resolve_fleet_host("http://192.168.25.50:11434")
-    assert url == "http://192.168.25.50:11434"
-    assert hostname == "node1"
-
-
-def test_resolve_fleet_host_dns_failure(monkeypatch) -> None:
-    _resolve_fleet_host.cache_clear()
-    def _raise(ip: str) -> None:
-        raise socket.herror("not found")
-    monkeypatch.setattr(socket, "gethostbyaddr", _raise)
-    url, hostname = _resolve_fleet_host("http://192.168.25.50:11434")
-    assert url == "http://192.168.25.50:11434"
-    assert hostname is None
-
-
-def test_resolve_fleet_host_strips_port_before_lookup(monkeypatch) -> None:
-    _resolve_fleet_host.cache_clear()
-    seen: list[str] = []
-    monkeypatch.setattr(socket, "gethostbyaddr", lambda ip: seen.append(ip) or ("node1", [], []))
-    _resolve_fleet_host("http://192.168.25.50:11434")
-    assert seen == ["192.168.25.50"]
-
-
-def test_resolve_fleet_host_returns_url_unchanged(monkeypatch) -> None:
-    _resolve_fleet_host.cache_clear()
-    monkeypatch.setattr(socket, "gethostbyaddr", lambda ip: ("gateway", [], []))
-    url, _ = _resolve_fleet_host("http://192.0.2.1:11434")
-    assert url == "http://192.0.2.1:11434"
